Mimic folder paths resolution on the client side
Previously, if one went to this URL "//toto", the static file "/toto" was
rendered by my hosting provider (netlify). However, once the client side
rendered the app, it didn't find "//toto" since there is no route
registered at it. Hence, it rendered a 404 page.

To solve this issue, the router now checks if the URL is correctly formed.
If not, it redirects to a correclty formed path.

Thank you @GillesDebunne